requirements.txt
asgiref==3.2.7
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.5
django-crispy-forms==1.9.0
django-import-export==2.0.2
django-taggit==1.2.0
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
psycopg2==2.8.5
pytz==2019.3
PyYAML==5.3.1
sqlparse==0.3.1
static3==0.7.0
tablib==1.1.0
wfastcgi==3.0.0
xlrd==1.2.0
xlwt==1.3.0
